Document-Analysis-System/Services/tokenizer-service/requirements.txt
pika
minio
python-docx
PyPDF2
